+++ hybrid_core/hybrid_core/example/android/app/src/main/kotlin/dev/hebei/hybrid_core_example/MainActivity.kt
@@ -1,5 +1,0 @@
-package dev.hebei.hybrid_core_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity()