+++ hybrid_os/hybrid_os_darwin/lib/src/ios.dart
@@ -1,0 +1,9 @@
+// ignore_for_file: camel_case_types
+
+import 'darwin.dart';
+
+/// iOSPlatform.
+final class iOSPlatform extends DarwinPlatform implements iOS {}
+
+/// iOS.
+abstract interface class iOS implements Darwin {}