+++ hybrid_os/hybrid_os_darwin/lib/src/macos.dart
@@ -1,0 +1,9 @@
+// ignore_for_file: camel_case_types
+
+import 'darwin.dart';
+
+/// macOSPlatform.
+final class macOSPlatform extends DarwinPlatform implements macOS {}
+
+/// macOS.
+abstract interface class macOS implements Darwin {}