+++ hybrid_os/hybrid_os_linux/lib/src/linux.dart
@@ -1,0 +1,7 @@
+import 'package:hybrid_os_platform_interface/hybrid_os_platform_interface.dart';
+
+/// LinuxPlatform.
+final class LinuxPlatform extends OSPlatform implements Linux {}
+
+/// Linux.
+abstract interface class Linux implements OS {}