+++ hybrid_os/hybrid_os_linux/lib/src/os.dart
@@ -1,0 +1,11 @@
+import 'package:hybrid_os_platform_interface/hybrid_os_platform_interface.dart';
+
+import 'linux.dart';
+
+/// HybridOSLinuxPlugin.
+abstract final class HybridOSLinuxPlugin {
+  /// Registers the [HybridOSLinuxPlugin].
+  static void registerWith() {
+    OSPlatform.instance = LinuxPlatform();
+  }
+}