+++ hybrid_os/hybrid_os/CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Add `OS` API.
+* Add `Android` and `AndroidSDKVersions` APIs.
+* Add `Darwin`, `DarwinVersion`, `iOS` and `macOS` APIs.
+* Add `Windows`, `WindowsVersion` and `WindowsType` APIs.
+* Add `Linux` API.